requirements.txt
Flask
akshare
redis
Flask-SQLAlchemy
psycopg2-binary
google-generativeai
python-dotenv
Flask-CORS
Flask-JWT-Extended
Flask-Migrate
